Reviewer comments and suggestions incorporated.
