+++ cmd/kpod/common.go
@@ -136,11 +136,11 @@ func validateFlags(c *cli.Context, flags []cli.Flag) error {
 
 // Common flags shared between commands
 var createFlags = []cli.Flag{
-	cli.StringSliceFlag{ //
+	cli.StringSliceFlag{
 		Name:  "add-host",
 		Usage: "Add a custom host-to-IP mapping (host:ip) (default [])",
 	},
-	cli.StringSliceFlag{ //
+	cli.StringSliceFlag{
 		Name:  "attach, a",
 		Usage: "Attach to STDIN, STDOUT or STDERR (default [])",
 	},
@@ -164,12 +164,8 @@ var createFlags = []cli.Flag{
 		Name:  "cgroup-parent",
 		Usage: "Optional parent cgroup for the container",
 	},
-	cli.Int64Flag{
-		Name:  "cpu-count",
-		Usage: "Limit the number of CPUs available for execution by the container.",
-	},
-	cli.StringFlag{ //
-		Name:  "cid-file",
+	cli.StringFlag{
+		Name:  "cidfile",
 		Usage: "Write the container ID to the file",
 	},
 	cli.Uint64Flag{
@@ -208,7 +204,7 @@ var createFlags = []cli.Flag{
 		Name:  "detach, d",
 		Usage: "Run container in background and print container ID",
 	},
-	cli.StringFlag{ //
+	cli.StringFlag{
 		Name:  "detach-keys",
 		Usage: "Override the key sequence for detaching a container. Format is a single character `[a-Z]` or `ctrl-<value>` where `<value>` is one of: `a-z`, `@`, `^`, `[`, `,` or `_`",
 	},
@@ -253,7 +249,7 @@ var createFlags = []cli.Flag{
 		Usage: "Set environment variables in container",
 	},
 	cli.StringSliceFlag{
-		Name:  "env-file", //
+		Name:  "env-file",
 		Usage: "Read in a file of environment variables",
 	},
 	cli.StringSliceFlag{
@@ -280,7 +276,7 @@ var createFlags = []cli.Flag{
 		Name:  "ip6",
 		Usage: "Container IPv6 address (e.g. 2001:db8::1b99)",
 	},
-	cli.StringFlag{ //
+	cli.StringFlag{
 		Name:  "ipc",
 		Usage: "IPC Namespace to use",
 	},
@@ -292,7 +288,7 @@ var createFlags = []cli.Flag{
 		Name:  "label",
 		Usage: "Set metadata on container (default [])",
 	},
-	cli.StringSliceFlag{ //
+	cli.StringSliceFlag{
 		Name:  "label-file",
 		Usage: "Read in a line delimited file of labels (default [])",
 	},
@@ -344,7 +340,7 @@ var createFlags = []cli.Flag{
 		Name:  "network-alias",
 		Usage: "Add network-scoped alias for the container (default [])",
 	},
-	cli.BoolFlag{ //
+	cli.BoolFlag{
 		Name:  "oom-kill-disable",
 		Usage: "Disable OOM Killer",
 	},
