+++ cmd/kpod/create.go
@@ -2,19 +2,13 @@ package main
 
 import (
 	"fmt"
-
-	spec "github.com/opencontainers/runtime-spec/specs-go"
-	"github.com/pkg/errors"
-	"github.com/urfave/cli"
-	pb "k8s.io/kubernetes/pkg/kubelet/apis/cri/v1alpha1/runtime"
-	"strings"
+	"strconv"
 
 	"github.com/docker/go-units"
 	"github.com/kubernetes-incubator/cri-o/libpod"
-	ann "github.com/kubernetes-incubator/cri-o/pkg/annotations"
-	"github.com/sirupsen/logrus"
-	"golang.org/x/sys/unix"
-	"strconv"
+	"github.com/pkg/errors"
+	"github.com/urfave/cli"
+	pb "k8s.io/kubernetes/pkg/kubelet/apis/cri/v1alpha1/runtime"
 )
 
 type mountType string
@@ -24,11 +18,9 @@ const (
 	// TypeBind is the type for mounting host dir
 	TypeBind mountType = "bind"
 	// TypeVolume is the type for remote storage volumes
-	TypeVolume mountType = "volume"
+	// TypeVolume mountType = "volume"  // re-enable upon use
 	// TypeTmpfs is the type for mounting tmpfs
 	TypeTmpfs mountType = "tmpfs"
-	// TypeNamedPipe is the type for mounting Windows named pipes
-	TypeNamedPipe mountType = "npipe"
 )
 
 var (
@@ -138,8 +130,8 @@ func createCmd(c *cli.Context) error {
 	if err := validateFlags(c, createFlags); err != nil {
 		return err
 	}
-	//runtime, err := getRuntime(c)
-	runtime, err := libpod.NewRuntime()
+
+	runtime, err := getRuntime(c)
 	if err != nil {
 		return errors.Wrapf(err, "error creating libpod runtime")
 	}
@@ -167,12 +159,17 @@ func createCmd(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	fmt.Println(imageName)
 	imageID, err := createImage.GetImageID()
 	if err != nil {
 		return err
 	}
-	ctr, err := runtime.NewContainer(runtimeSpec, libpod.WithRootFSFromImage(imageID, imageName, false))
+	options, err := createConfig.GetContainerCreateOptions(c)
+	if err != nil {
+		return errors.Wrapf(err, "unable to parse new container options")
+	}
+	// Gather up the options for NewContainer which consist of With... funcs
+	options = append(options, libpod.WithRootFSFromImage(imageID, imageName, false))
+	ctr, err := runtime.NewContainer(runtimeSpec, options...)
 	if err != nil {
 		return err
 	}
@@ -181,28 +178,13 @@ func createCmd(c *cli.Context) error {
 		return err
 	}
 
-	if c.String("cid-file") != "" {
-		libpod.WriteFile(ctr.ID(), c.String("cid-file"))
-		return nil
+	if c.String("cidfile") != "" {
+		libpod.WriteFile(ctr.ID(), c.String("cidfile"))
+	} else {
+		fmt.Printf("%s\n", ctr.ID())
 	}
-	fmt.Printf("%s\n", ctr.ID())
 
 	return nil
-}
-
-/* The following funcs should land in parse.go */
-//
-//
-func stringSlicetoUint32Slice(inputSlice []string) ([]uint32, error) {
-	var outputSlice []uint32
-	for _, v := range inputSlice {
-		u, err := strconv.ParseUint(v, 10, 32)
-		if err != nil {
-			return outputSlice, err
-		}
-		outputSlice = append(outputSlice, uint32(u))
-	}
-	return outputSlice, nil
 }
 
 // Parses CLI options related to container creation into a config which can be
@@ -211,51 +193,32 @@ func parseCreateOpts(c *cli.Context, runtime *libpod.Runtime) (*createConfig, er
 	var command []string
 	var memoryLimit, memoryReservation, memorySwap, memoryKernel int64
 	var blkioWeight uint16
-	var env []string
-	var labelValues []string
 	var uid, gid uint32
-	sysctl := make(map[string]string)
-	labels := make(map[string]string)
 
 	image := c.Args()[0]
 
 	if len(c.Args()) < 1 {
-		return nil, errors.Errorf("you just provide an image name")
+		return nil, errors.Errorf("image name or ID is required")
 	}
 	if len(c.Args()) > 1 {
 		command = c.Args()[1:]
 	}
 
 	// LABEL VARIABLES
-	// TODO where should labels be verified to be x=y format
-	labelValues, labelErr := readKVStrings(c.StringSlice("label-file"), c.StringSlice("label"))
-	if labelErr != nil {
-		return &createConfig{}, errors.Wrapf(labelErr, "unable to process labels from --label and label-file")
+	labels, err := getAllLabels(c)
+	if err != nil {
+		return &createConfig{}, errors.Wrapf(err, "unable to process labels")
 	}
-	// Process KEY=VALUE stringslice in string map for WithLabels func
-	if len(labelValues) > 0 {
-		for _, i := range labelValues {
-			spliti := strings.Split(i, "=")
-			labels[spliti[0]] = spliti[1]
-		}
-	}
-
 	// ENVIRONMENT VARIABLES
 	// TODO where should env variables be verified to be x=y format
-	env, err := readKVStrings(c.StringSlice("env-file"), c.StringSlice("env"))
+	env, err := getAllEnvironmentVariables(c)
 	if err != nil {
-		return &createConfig{}, errors.Wrapf(err, "unable to process variables from --env and --env-file")
-	}
-	// Add default environment variables if nothing defined
-	if len(env) == 0 {
-		env = append(env, defaultEnvVariables...)
+		return &createConfig{}, errors.Wrapf(err, "unable to process environment variables")
 	}
 
-	if len(c.StringSlice("sysctl")) > 0 {
-		for _, inputSysctl := range c.StringSlice("sysctl") {
-			values := strings.Split(inputSysctl, "=")
-			sysctl[values[0]] = values[1]
-		}
+	sysctl, err := convertStringSliceToMap(c.StringSlice("sysctl"), "=")
+	if err != nil {
+		return &createConfig{}, errors.Wrapf(err, "sysctl values must be in the form of KEY=VALUE")
 	}
 
 	groupAdd, err := stringSlicetoUint32Slice(c.StringSlice("group-add"))
@@ -338,10 +301,9 @@ func parseCreateOpts(c *cli.Context, runtime *libpod.Runtime) (*createConfig, er
 		publishAll:     c.Bool("publish-all"),
 		readOnlyRootfs: c.Bool("read-only"),
 		resources: createResourceConfig{
-			blkioWeight: blkioWeight,
-			blkioDevice: c.StringSlice("blkio-weight-device"),
-			cpuShares:   c.Uint64("cpu-shares"),
-			//cpuCount:          c.Int64("cpu-count"),
+			blkioWeight:       blkioWeight,
+			blkioDevice:       c.StringSlice("blkio-weight-device"),
+			cpuShares:         c.Uint64("cpu-shares"),
 			cpuPeriod:         c.Uint64("cpu-period"),
 			cpusetCpus:        c.String("cpu-period"),
 			cpusetMems:        c.String("cpuset-mems"),
@@ -354,6 +316,7 @@ func parseCreateOpts(c *cli.Context, runtime *libpod.Runtime) (*createConfig, er
 			deviceWriteBps:    c.StringSlice("device-write-bps"),
 			deviceWriteIops:   c.StringSlice("device-write-iops"),
 			disableOomKiller:  c.Bool("oom-kill-disable"),
+			shmSize:           c.String("shm-size"),
 			memory:            memoryLimit,
 			memoryReservation: memoryReservation,
 			memorySwap:        memorySwap,
@@ -366,534 +329,19 @@ func parseCreateOpts(c *cli.Context, runtime *libpod.Runtime) (*createConfig, er
 		},
 		rm:           c.Bool("rm"),
 		securityOpts: c.StringSlice("security-opt"),
-		//shmSize: c.String("shm-size"),
-		sigProxy:    c.Bool("sig-proxy"),
-		stopSignal:  c.String("stop-signal"),
-		stopTimeout: c.Int64("stop-timeout"),
-		storageOpts: c.StringSlice("storage-opt"),
-		sysctl:      sysctl,
-		tmpfs:       c.StringSlice("tmpfs"),
-		tty:         c.Bool("tty"), //
-		user:        uid,
-		group:       gid,
-		//userns: c.String("userns"),
-		volumes:     c.StringSlice("volume"),
-		volumesFrom: c.StringSlice("volumes-from"),
-		workDir:     c.String("workdir"),
+		sigProxy:     c.Bool("sig-proxy"),
+		stopSignal:   c.String("stop-signal"),
+		stopTimeout:  c.Int64("stop-timeout"),
+		storageOpts:  c.StringSlice("storage-opt"),
+		sysctl:       sysctl,
+		tmpfs:        c.StringSlice("tmpfs"),
+		tty:          c.Bool("tty"),
+		user:         uid,
+		group:        gid,
+		volumes:      c.StringSlice("volume"),
+		volumesFrom:  c.StringSlice("volumes-from"),
+		workDir:      c.String("workdir"),
 	}
 
 	return config, nil
-}
-
-// Parses information needed to create a container into an OCI runtime spec
-func createConfigToOCISpec(config *createConfig) (*spec.Spec, error) {
-
-	//blkio, err := config.CreateBlockIO()
-	//if err != nil {
-	//	return &spec.Spec{}, err
-	//}
-
-	spec := config.GetDefaultLinuxSpec()
-	spec.Process.Cwd = config.workDir
-	spec.Process.Args = config.command
-
-	if config.tty {
-		spec.Process.Terminal = config.tty
-	}
-
-	if config.user != 0 {
-		// User and Group must go together
-		spec.Process.User.UID = config.user
-		spec.Process.User.GID = config.group
-	}
-	if len(config.groupAdd) > 0 {
-		spec.Process.User.AdditionalGids = config.groupAdd
-	}
-	if len(config.env) > 0 {
-		spec.Process.Env = config.env
-	}
-	//TODO
-	// Need examples of capacity additions so I can load that properly
-
-	if config.readOnlyRootfs {
-		spec.Root.Readonly = config.readOnlyRootfs
-	}
-
-	if config.hostname != "" {
-		spec.Hostname = config.hostname
-	}
-
-	// BIND MOUNTS
-	if len(config.volumes) > 0 {
-		spec.Mounts = append(spec.Mounts, config.GetVolumeMounts()...)
-	}
-	// TMPFS MOUNTS
-	if len(config.tmpfs) > 0 {
-		spec.Mounts = append(spec.Mounts, config.GetTmpfsMounts()...)
-	}
-
-	// RESOURCES - MEMORY
-	if len(config.sysctl) > 0 {
-		spec.Linux.Sysctl = config.sysctl
-	}
-	if config.resources.memory != 0 {
-		spec.Linux.Resources.Memory.Limit = &config.resources.memory
-	}
-	if config.resources.memoryReservation != 0 {
-		spec.Linux.Resources.Memory.Reservation = &config.resources.memoryReservation
-	}
-	if config.resources.memorySwap != 0 {
-		spec.Linux.Resources.Memory.Swap = &config.resources.memorySwap
-	}
-	if config.resources.kernelMemory != 0 {
-		spec.Linux.Resources.Memory.Kernel = &config.resources.kernelMemory
-	}
-	if config.resources.memorySwapiness != 0 {
-		spec.Linux.Resources.Memory.Swappiness = &config.resources.memorySwapiness
-	}
-	if config.resources.disableOomKiller {
-		spec.Linux.Resources.Memory.DisableOOMKiller = &config.resources.disableOomKiller
-	}
-
-	// RESOURCES - CPU
-
-	if config.resources.cpuShares != 0 {
-		spec.Linux.Resources.CPU.Shares = &config.resources.cpuShares
-	}
-	if config.resources.cpuQuota != 0 {
-		spec.Linux.Resources.CPU.Quota = &config.resources.cpuQuota
-	}
-	if config.resources.cpuPeriod != 0 {
-		spec.Linux.Resources.CPU.Period = &config.resources.cpuPeriod
-	}
-	if config.resources.cpuRtRuntime != 0 {
-		spec.Linux.Resources.CPU.RealtimeRuntime = &config.resources.cpuRtRuntime
-	}
-	if config.resources.cpuRtPeriod != 0 {
-		spec.Linux.Resources.CPU.RealtimePeriod = &config.resources.cpuRtPeriod
-	}
-	if config.resources.cpus != "" {
-		spec.Linux.Resources.CPU.Cpus = config.resources.cpus
-	}
-	if config.resources.cpusetMems != "" {
-		spec.Linux.Resources.CPU.Mems = config.resources.cpusetMems
-	}
-
-	// RESOURCES - PIDS
-	if config.resources.pidsLimit != 0 {
-		spec.Linux.Resources.Pids.Limit = config.resources.pidsLimit
-	}
-
-	/*
-				Capabilities: &spec.LinuxCapabilities{
-				// Rlimits []PosixRlimit // Where does this come from
-				// Type string
-				// Hard uint64
-				// Limit uint64
-				// NoNewPrivileges bool // No user input for this
-				// ApparmorProfile string // No user input for this
-				OOMScoreAdj: &config.resources.oomScoreAdj,
-				// Selinuxlabel
-			},
-			Hooks: &spec.Hooks{},
-			//Annotations
-				Resources: &spec.LinuxResources{
-					Devices: config.GetDefaultDevices(),
-					BlockIO: &blkio,
-					//HugepageLimits:
-					Network: &spec.LinuxNetwork{
-					// ClassID *uint32
-					// Priorites []LinuxInterfacePriority
-					},
-				},
-				//CgroupsPath:
-				//Namespaces: []LinuxNamespace
-				//Devices
-				Seccomp: &spec.LinuxSeccomp{
-				// DefaultAction:
-				// Architectures
-				// Syscalls:
-				},
-				// RootfsPropagation
-				// MaskedPaths
-				// ReadonlyPaths:
-				// MountLabel
-				// IntelRdt
-			},
-		}
-	*/
-	return &spec, nil
-}
-
-func getStatFromPath(path string) unix.Stat_t {
-	s := unix.Stat_t{}
-	_ = unix.Stat(path, &s)
-	return s
-}
-
-func makeThrottleArray(throttleInput []string) ([]spec.LinuxThrottleDevice, error) {
-	var ltds []spec.LinuxThrottleDevice
-	for _, i := range throttleInput {
-		t, err := validateBpsDevice(i)
-		if err != nil {
-			return []spec.LinuxThrottleDevice{}, err
-		}
-		ltd := spec.LinuxThrottleDevice{}
-		ltd.Rate = t.rate
-		ltdStat := getStatFromPath(t.path)
-		ltd.Major = int64(unix.Major(ltdStat.Rdev))
-		ltd.Minor = int64(unix.Major(ltdStat.Rdev))
-		ltds = append(ltds, ltd)
-	}
-	return ltds, nil
-
-}
-
-func (c *createConfig) CreateBlockIO() (spec.LinuxBlockIO, error) {
-	bio := spec.LinuxBlockIO{}
-	bio.Weight = &c.resources.blkioWeight
-	if len(c.resources.blkioDevice) > 0 {
-		var lwds []spec.LinuxWeightDevice
-		for _, i := range c.resources.blkioDevice {
-			wd, err := validateweightDevice(i)
-			if err != nil {
-				return bio, errors.Wrapf(err, "invalid values for blkio-weight-device")
-			}
-			wdStat := getStatFromPath(wd.path)
-			lwd := spec.LinuxWeightDevice{
-				Weight: &wd.weight,
-			}
-			lwd.Major = int64(unix.Major(wdStat.Rdev))
-			lwd.Minor = int64(unix.Minor(wdStat.Rdev))
-			lwds = append(lwds, lwd)
-		}
-	}
-	if len(c.resources.deviceReadBps) > 0 {
-		readBps, err := makeThrottleArray(c.resources.deviceReadBps)
-		if err != nil {
-			return bio, err
-		}
-		bio.ThrottleReadBpsDevice = readBps
-	}
-	if len(c.resources.deviceWriteBps) > 0 {
-		writeBpds, err := makeThrottleArray(c.resources.deviceWriteBps)
-		if err != nil {
-			return bio, err
-		}
-		bio.ThrottleWriteBpsDevice = writeBpds
-	}
-	if len(c.resources.deviceReadIops) > 0 {
-		readIops, err := makeThrottleArray(c.resources.deviceReadIops)
-		if err != nil {
-			return bio, err
-		}
-		bio.ThrottleReadIOPSDevice = readIops
-	}
-	if len(c.resources.deviceWriteIops) > 0 {
-		writeIops, err := makeThrottleArray(c.resources.deviceWriteIops)
-		if err != nil {
-			return bio, err
-		}
-		bio.ThrottleWriteIOPSDevice = writeIops
-	}
-
-	return bio, nil
-}
-
-func (c *createConfig) GetDefaultMounts() []spec.Mount {
-	return []spec.Mount{
-		{
-			Destination: "/proc",
-			Type:        "proc",
-			Source:      "proc",
-			Options:     []string{"nosuid", "noexec", "nodev"},
-		},
-		{
-			Destination: "/dev",
-			Type:        "tmpfs",
-			Source:      "tmpfs",
-			Options:     []string{"nosuid", "strictatime", "mode=755", "size=65536k"},
-		},
-		{
-			Destination: "/dev/pts",
-			Type:        "devpts",
-			Source:      "devpts",
-			Options:     []string{"nosuid", "noexec", "newinstance", "ptmxmode=0666", "mode=0620", "gid=5"},
-		},
-		{
-			Destination: "/sys",
-			Type:        "sysfs",
-			Source:      "sysfs",
-			Options:     []string{"nosuid", "noexec", "nodev", "ro"},
-		},
-		{
-			Destination: "/sys/fs/cgroup",
-			Type:        "cgroup",
-			Source:      "cgroup",
-			Options:     []string{"ro", "nosuid", "noexec", "nodev"},
-		},
-		{
-			Destination: "/dev/mqueue",
-			Type:        "mqueue",
-			Source:      "mqueue",
-			Options:     []string{"nosuid", "noexec", "nodev"},
-		},
-		{
-			Destination: "/dev/shm",
-			Type:        "tmpfs",
-			Source:      "shm",
-			Options:     []string{"nosuid", "noexec", "nodev", "mode=1777"},
-		},
-	}
-}
-func iPtr(i int64) *int64 { return &i }
-
-func (c *createConfig) GetDefaultDevices() []spec.LinuxDeviceCgroup {
-	return []spec.LinuxDeviceCgroup{
-		{
-			Allow:  false,
-			Access: "rwm",
-		},
-		{
-			Allow:  true,
-			Type:   "c",
-			Major:  iPtr(1),
-			Minor:  iPtr(5),
-			Access: "rwm",
-		},
-		{
-			Allow:  true,
-			Type:   "c",
-			Major:  iPtr(1),
-			Minor:  iPtr(3),
-			Access: "rwm",
-		},
-		{
-			Allow:  true,
-			Type:   "c",
-			Major:  iPtr(1),
-			Minor:  iPtr(9),
-			Access: "rwm",
-		},
-		{
-			Allow:  true,
-			Type:   "c",
-			Major:  iPtr(1),
-			Minor:  iPtr(8),
-			Access: "rwm",
-		},
-		{
-			Allow:  true,
-			Type:   "c",
-			Major:  iPtr(5),
-			Minor:  iPtr(0),
-			Access: "rwm",
-		},
-		{
-			Allow:  true,
-			Type:   "c",
-			Major:  iPtr(5),
-			Minor:  iPtr(1),
-			Access: "rwm",
-		},
-		{
-			Allow:  false,
-			Type:   "c",
-			Major:  iPtr(10),
-			Minor:  iPtr(229),
-			Access: "rwm",
-		},
-	}
-}
-
-func defaultCapabilities() []string {
-	return []string{
-		"CAP_CHOWN",
-		"CAP_DAC_OVERRIDE",
-		"CAP_FSETID",
-		"CAP_FOWNER",
-		"CAP_MKNOD",
-		"CAP_NET_RAW",
-		"CAP_SETGID",
-		"CAP_SETUID",
-		"CAP_SETFCAP",
-		"CAP_SETPCAP",
-		"CAP_NET_BIND_SERVICE",
-		"CAP_SYS_CHROOT",
-		"CAP_KILL",
-		"CAP_AUDIT_WRITE",
-	}
-}
-
-func (c *createConfig) GetDefaultLinuxSpec() spec.Spec {
-	s := spec.Spec{
-		Version: spec.Version,
-		Root:    &spec.Root{},
-	}
-	s.Annotations = c.GetAnnotations()
-	s.Mounts = c.GetDefaultMounts()
-	s.Process = &spec.Process{
-		Capabilities: &spec.LinuxCapabilities{
-			Bounding:    defaultCapabilities(),
-			Permitted:   defaultCapabilities(),
-			Inheritable: defaultCapabilities(),
-			Effective:   defaultCapabilities(),
-		},
-	}
-	s.Linux = &spec.Linux{
-		MaskedPaths: []string{
-			"/proc/kcore",
-			"/proc/latency_stats",
-			"/proc/timer_list",
-			"/proc/timer_stats",
-			"/proc/sched_debug",
-		},
-		ReadonlyPaths: []string{
-			"/proc/asound",
-			"/proc/bus",
-			"/proc/fs",
-			"/proc/irq",
-			"/proc/sys",
-			"/proc/sysrq-trigger",
-		},
-		Namespaces: []spec.LinuxNamespace{
-			{Type: "mount"},
-			{Type: "network"},
-			{Type: "uts"},
-			{Type: "pid"},
-			{Type: "ipc"},
-		},
-		Devices: []spec.LinuxDevice{},
-		Resources: &spec.LinuxResources{
-			Devices: c.GetDefaultDevices(),
-		},
-	}
-
-	return s
-}
-
-func (c *createConfig) GetAnnotations() map[string]string {
-	a := getDefaultAnnotations()
-	// TODO
-	// Which annotations do we want added by default
-	if c.tty {
-		a["io.kubernetes.cri-o.TTY"] = "true"
-	}
-	return a
-}
-
-func getDefaultAnnotations() map[string]string {
-	var a map[string]string
-	a = make(map[string]string)
-	a[ann.Annotations] = ""
-	a[ann.ContainerID] = ""
-	a[ann.ContainerName] = ""
-	a[ann.ContainerType] = ""
-	a[ann.Created] = ""
-	a[ann.HostName] = ""
-	a[ann.IP] = ""
-	a[ann.Image] = ""
-	a[ann.ImageName] = ""
-	a[ann.ImageRef] = ""
-	a[ann.KubeName] = ""
-	a[ann.Labels] = ""
-	a[ann.LogPath] = ""
-	a[ann.Metadata] = ""
-	a[ann.Name] = ""
-	a[ann.PrivilegedRuntime] = ""
-	a[ann.ResolvPath] = ""
-	a[ann.HostnamePath] = ""
-	a[ann.SandboxID] = ""
-	a[ann.SandboxName] = ""
-	a[ann.ShmPath] = ""
-	a[ann.MountPoint] = ""
-	a[ann.TrustedSandbox] = ""
-	a[ann.TTY] = "false"
-	a[ann.Stdin] = ""
-	a[ann.StdinOnce] = ""
-	a[ann.Volumes] = ""
-
-	return a
-}
-
-//GetTmpfsMounts takes user provided input for bind mounts and creates Mount structs
-func (c *createConfig) GetVolumeMounts() []spec.Mount {
-	var m []spec.Mount
-	var options []string
-	for _, i := range c.volumes {
-		spliti := strings.Split(i, ":")
-		if len(spliti) > 2 {
-			options = strings.Split(spliti[2], ",")
-		}
-		// always add rbind bc mount ignores the bind filesystem when mounting
-		options = append(options, "rbind")
-		m = append(m, spec.Mount{
-			Destination: spliti[1],
-			Type:        string(TypeBind),
-			Source:      spliti[0],
-			Options:     options,
-		})
-	}
-	return m
-}
-
-//GetTmpfsMounts takes user provided input for tmpfs mounts and creates Mount structs
-func (c *createConfig) GetTmpfsMounts() []spec.Mount {
-	var m []spec.Mount
-	for _, i := range c.tmpfs {
-		// Default options if nothing passed
-		options := []string{"rw", "noexec", "nosuid", "nodev", "size=65536k"}
-		spliti := strings.Split(i, ":")
-		destPath := spliti[0]
-		if len(spliti) > 1 {
-			options = strings.Split(spliti[1], ",")
-		}
-		m = append(m, spec.Mount{
-			Destination: destPath,
-			Type:        string(TypeTmpfs),
-			Options:     options,
-		})
-	}
-	return m
-}
-
-func (c *createConfig) GetContainerCreateOptions(cli *cli.Context) ([]libpod.CtrCreateOption, error) {
-	/*
-		WithStorageConfig
-		WithImageConfig
-		WithSignaturePolicy
-		WithOCIRuntime
-		WithConmonPath
-		WithConmonEnv
-		WithCgroupManager
-		WithStaticDir
-		WithTmpDir
-		WithSELinux
-		WithPidsLimit // dont need
-		WithMaxLogSize
-		WithNoPivotRoot
-		WithRootFSFromPath
-		WithRootFSFromImage
-		WithStdin // done
-		WithSharedNamespaces
-		WithLabels //done
-		WithAnnotations // dont need
-		WithName // done
-		WithStopSignal
-		WithPodName
-	*/
-	var options []libpod.CtrCreateOption
-
-	// Uncomment after talking to mheon about unimplemented funcs
-	// options = append(options, libpod.WithLabels(c.labels))
-
-	if c.interactive {
-		options = append(options, libpod.WithStdin())
-	}
-	if c.name != "" {
-		logrus.Info("appending name %s", c.name)
-		options = append(options, libpod.WithName(c.name))
-	}
-
-	return options, nil
 }
