+++ libpod/storage.go
@@ -57,7 +57,7 @@ func (metadata *RuntimeContainerMetadata) SetMountLabel(mountLabel string) {
 }
 
 // CreateContainerStorage creates the storage end of things.  We already have the container spec created
-// TO-DO We should be passing in an KpodImage object in the future.
+// TO-DO We should be passing in an Image object in the future.
 func (r *storageService) CreateContainerStorage(systemContext *types.SystemContext, imageName, imageID, containerName, containerID, mountLabel string) (ContainerInfo, error) {
 	var ref types.ImageReference
 	if imageName == "" && imageID == "" {
